count 관련 버그 수정
exit 펑션에서 worker 인스턴스 체크 로직 추가

+++ ZU4A_IMP/U4A_LIB/v1000/resources/u4a/m/CountdownTimer-dbg.js
@@ -124,6 +124,7 @@ sap.ui.define("u4a.m.CountdownTimer",[
 
             if(this._oWorker instanceof Worker){
 				this._oWorker.terminate();
+				this._oWorker = null;
             }
 
             this._oWorker = new Worker('/zu4a_imp/publish/CommonJS/workers/CountdownTimerWorker.js');
@@ -147,9 +148,11 @@ sap.ui.define("u4a.m.CountdownTimer",[
         }, // end of onAfterRendering
 		
 		exit : function(){
-			        
-			this._oWorker.terminate();
-			this._oWorker = null;
+			
+			if(this._oWorker){
+				this._oWorker.terminate();
+				this._oWorker = null;				
+			}			
 			
 		}
